requirements.txt
playwright
supabase
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
requests
instagrapi
Pillow
